toInt 수정완료 했습니다.
